+++ app/Http/Controllers/TagsController.php
@@ -93,12 +93,14 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash('errors', 'Tag can not be deleted because it is associated to some posts');
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
 
         return redirect(route('tags.index'));
     }
-
-    
 }
